merge.

+++ app/Http/Controllers/UserController.php
@@ -11,28 +11,31 @@ use Request;
 use Redirect;
 use Session;
 use URL;
+use Password;
 use Utils;
 use Validator;
-use Illuminate\Auth\Passwords\TokenRepositoryInterface;
 use App\Models\User;
 use App\Http\Requests;
 use App\Ninja\Repositories\AccountRepository;
 use App\Ninja\Mailers\ContactMailer;
 use App\Ninja\Mailers\UserMailer;
+use App\Services\UserService;
 
 class UserController extends BaseController
 {
     protected $accountRepo;
     protected $contactMailer;
     protected $userMailer;
+    protected $userService;
 
-    public function __construct(AccountRepository $accountRepo, ContactMailer $contactMailer, UserMailer $userMailer)
+    public function __construct(AccountRepository $accountRepo, ContactMailer $contactMailer, UserMailer $userMailer, UserService $userService)
     {
-        parent::__construct();
+        //parent::__construct();
 
         $this->accountRepo = $accountRepo;
         $this->contactMailer = $contactMailer;
         $this->userMailer = $userMailer;
+        $this->userService = $userService;
     }
 
     public function index()
@@ -42,47 +45,7 @@ class UserController extends BaseController
 
     public function getDatatable()
     {
-        $query = DB::table('users')
-                  ->where('users.account_id', '=', Auth::user()->account_id);
-
-        if (!Session::get('show_trash:user')) {
-            $query->where('users.deleted_at', '=', null);
-        }
-
-        $query->where('users.public_id', '>', 0)
-              ->select('users.public_id', 'users.first_name', 'users.last_name', 'users.email', 'users.confirmed', 'users.public_id', 'users.deleted_at');
-
-        return Datatable::query($query)
-        ->addColumn('first_name', function ($model) { return link_to('users/'.$model->public_id.'/edit', $model->first_name.' '.$model->last_name); })
-        ->addColumn('email', function ($model) { return $model->email; })
-        ->addColumn('confirmed', function ($model) { return $model->deleted_at ? trans('texts.deleted') : ($model->confirmed ? trans('texts.active') : trans('texts.pending')); })
-        ->addColumn('dropdown', function ($model) {
-          $actions = '<div class="btn-group tr-action" style="visibility:hidden;">
-              <button type="button" class="btn btn-xs btn-default dropdown-toggle" data-toggle="dropdown">
-                '.trans('texts.select').' <span class="caret"></span>
-              </button>
-              <ul class="dropdown-menu" role="menu">';
-
-          if ($model->deleted_at) {
-              $actions .= '<li><a href="'.URL::to('restore_user/'.$model->public_id).'">'.uctrans('texts.restore_user').'</a></li>';
-          } else {
-              $actions .= '<li><a href="'.URL::to('users/'.$model->public_id).'/edit">'.uctrans('texts.edit_user').'</a></li>';
-
-              if (!$model->confirmed) {
-                  $actions .= '<li><a href="'.URL::to('send_confirmation/'.$model->public_id).'">'.uctrans('texts.send_invite').'</a></li>';
-              }
-
-              $actions .= '<li class="divider"></li>
-                <li><a href="javascript:deleteUser('.$model->public_id.')">'.uctrans('texts.delete_user').'</a></li>';
-          }
-
-           $actions .= '</ul>
-          </div>';
-
-          return $actions;
-        })
-        ->orderColumns(['first_name', 'email', 'confirmed'])
-        ->make();
+        return $this->userService->getDatatable(Auth::user()->account_id);
     }
 
     public function setTheme()
@@ -114,7 +77,6 @@ class UserController extends BaseController
             'user' => $user,
             'method' => 'PUT',
             'url' => 'users/'.$publicId,
-            'title' => trans('texts.edit_user'),
         ];
 
         return View::make('users.edit', $data);
@@ -139,7 +101,7 @@ class UserController extends BaseController
         if (!Auth::user()->registered) {
             Session::flash('error', trans('texts.register_to_add_user'));
             return Redirect::to('settings/' . ACCOUNT_USER_MANAGEMENT);
-        }        
+        }
         if (!Auth::user()->confirmed) {
             Session::flash('error', trans('texts.confirmation_required'));
             return Redirect::to('settings/' . ACCOUNT_USER_MANAGEMENT);
@@ -157,21 +119,28 @@ class UserController extends BaseController
           'user' => null,
           'method' => 'POST',
           'url' => 'users',
-          'title' => trans('texts.add_user'),
         ];
 
         return View::make('users.edit', $data);
     }
 
-    public function delete()
+    public function bulk()
     {
-        $userPublicId = Input::get('userPublicId');
+        $action = Input::get('bulk_action');
+        $id = Input::get('bulk_public_id');
+
         $user = User::where('account_id', '=', Auth::user()->account_id)
-                    ->where('public_id', '=', $userPublicId)->firstOrFail();
+                    ->where('public_id', '=', $id)
+                    ->withTrashed()
+                    ->firstOrFail();
 
-        $user->delete();
+        if ($action === 'archive') {
+            $user->delete();
+        } else {
+            $user->restore();
+        }
 
-        Session::flash('message', trans('texts.deleted_user'));
+        Session::flash('message', trans("texts.{$action}d_user"));
 
         return Redirect::to('settings/' . ACCOUNT_USER_MANAGEMENT);
     }
@@ -195,7 +164,7 @@ class UserController extends BaseController
      */
     public function save($userPublicId = false)
     {
-        if (Auth::user()->account->isPro()) {
+        if (Auth::user()->isPro() && ! Auth::user()->isTrial()) {
             $rules = [
                 'first_name' => 'required',
                 'last_name' => 'required',
@@ -221,6 +190,8 @@ class UserController extends BaseController
                 $user->last_name = trim(Input::get('last_name'));
                 $user->username = trim(Input::get('email'));
                 $user->email = trim(Input::get('email'));
+                $user->is_admin = boolval(Input::get('is_admin'));
+                $user->permissions = Input::get('permissions');
             } else {
                 $lastUser = User::withTrashed()->where('account_id', '=', Auth::user()->account_id)
                             ->orderBy('public_id', 'DESC')->first();
@@ -231,10 +202,12 @@ class UserController extends BaseController
                 $user->last_name = trim(Input::get('last_name'));
                 $user->username = trim(Input::get('email'));
                 $user->email = trim(Input::get('email'));
+                $user->is_admin = boolval(Input::get('is_admin'));
                 $user->registered = true;
                 $user->password = str_random(RANDOM_KEY_LENGTH);
                 $user->confirmation_code = str_random(RANDOM_KEY_LENGTH);
                 $user->public_id = $lastUser->public_id + 1;
+                $user->permissions = Input::get('permissions');
             }
 
             $user->save();
@@ -248,7 +221,7 @@ class UserController extends BaseController
 
             Session::flash('message', $message);
         }
-        
+
         return Redirect::to('settings/' . ACCOUNT_USER_MANAGEMENT);
     }
 
@@ -269,10 +242,10 @@ class UserController extends BaseController
      *
      * @param string $code
      */
-    public function confirm($code, TokenRepositoryInterface $tokenRepo)
+    public function confirm($code)
     {
         $user = User::where('confirmation_code', '=', $code)->get()->first();
-            
+
         if ($user) {
             $notice_msg = trans('texts.security.confirmation');
 
@@ -282,7 +255,7 @@ class UserController extends BaseController
 
             if ($user->public_id) {
                 //Auth::login($user);
-                $token = $tokenRepo->create($user);
+                $token = Password::getRepository()->create($user);
 
                 return Redirect::to("/password/reset/{$token}");
             } else {
@@ -323,7 +296,7 @@ class UserController extends BaseController
         return Redirect::to('/')->with('clearGuestKey', true);
     }
     */
-    
+
     public function changePassword()
     {
         // check the current password is correct
@@ -355,7 +328,7 @@ class UserController extends BaseController
         $oldUserId = Auth::user()->id;
         $referer = Request::header('referer');
         $account = $this->accountRepo->findUserAccounts($newUserId, $oldUserId);
-        
+
         if ($account) {
             if ($account->hasUserId($newUserId) && $account->hasUserId($oldUserId)) {
                 Auth::loginUsingId($newUserId);
@@ -366,7 +339,7 @@ class UserController extends BaseController
                 Session::put('_token', str_random(40));
             }
         }
-        
+
         return Redirect::to($referer);
     }
